Merge branch 'master' into michael-hennecke/2.3/daos-11247-memlock
Update DAOS version

+++ utils/run_test.sh
@@ -25,6 +25,49 @@ if [ -z "$DAOS_BASE" ]; then
     DAOS_BASE="."
 fi
 
+produce_output()
+{
+    echo "$2"
+
+    cname="run_test.${RUN_TEST_VALGRIND:-native}"
+    name="run_test"
+    fname="${DAOS_BASE}/test_results/test_run_test.sh.${RUN_TEST_VALGRIND:-native}.xml"
+
+    if [ "$1" -eq 0 ]; then
+       teststr="    <testcase classname=\"$cname\" name=\"$name\" />"
+    else
+       teststr="    <testcase classname=\"$cname\" name=\"$name\">
+      <failure type=\"format\">
+        <![CDATA[$2
+          ]]>
+      </failure>
+    </testcase>"
+    fi
+
+    cat > "${fname}" << EOF
+<?xml version="1.0" encoding="UTF-8" ?>
+<testsuites>
+  <testsuite tests="1" failures="$1" errors="0" skipped="0" >
+EOF
+echo "${teststr}" >> "${fname}"
+cat >> "${fname}" << EOF
+  </testsuite>
+</testsuites>
+EOF
+}
+
+run_test_filter()
+{
+    if [ -n "${RUN_TEST_FILTER}" ]; then
+        if ! [[ "${COMP} $*" =~ ${RUN_TEST_FILTER} ]]; then
+            echo "Skipping test: $*"
+            return 1
+        fi
+    fi
+
+    return 0
+}
+
 run_test()
 {
     local in="$*"
@@ -33,17 +76,15 @@ run_test()
     b="${b//;/-}"
     b="${b//\"/-}"
 
-    if [ -n "${RUN_TEST_FILTER}" ]; then
-        if ! [[ "$*" =~ ${RUN_TEST_FILTER} ]]; then
-            echo "Skipping test: $in"
-            return
-        fi
+    if ! run_test_filter "$*"; then
+        return
     fi
+
     export D_LOG_FILE="/tmp/daos_${b}-${log_num}.log"
-    echo "Running $* with log file: ${D_LOG_FILE}"
 
     export TNAME="${b}-${log_num}"
 
+    echo "Running $* with log file: ${D_LOG_FILE}"
     # We use grep to filter out any potential "SUCCESS! NO TEST FAILURES"
     #    messages as daos_post_build.sh will look for this and mark the tests
     #    as passed, which we don't want as we need to check all of the tests
@@ -60,6 +101,7 @@ run_test()
     ((log_num += 1))
 
     FILES=("${DAOS_BASE}"/test_results/*.xml)
+    sudo chown -R "$USER" "${FILES[@]}"
 
     "${SL_PREFIX}"/lib/daos/TESTING/ftest/scripts/post_process_xml.sh \
                                                                   "${COMP}" \
@@ -71,14 +113,10 @@ run_test()
 if [ -d "/mnt/daos" ]; then
     # shellcheck disable=SC1091
     source ./.build_vars.sh
-    if ! ${OLD_CI:-true}; then
-        # fix up paths so they are relative to $PWD since we might not
-        # be in the same path as the software was built
-        SL_PREFIX=$PWD/${SL_PREFIX/*\/install/install}
-    fi
 
     echo "Running Cmocka tests"
     mkdir -p "${DAOS_BASE}"/test_results/xml
+    chmod 777 "${DAOS_BASE}"/test_results/xml
 
     VALGRIND_CMD=""
     if [ -z "$RUN_TEST_VALGRIND" ]; then
@@ -104,22 +142,30 @@ if [ -d "/mnt/daos" ]; then
         run_test src/common/tests/btree.sh dyn perf ukey -s 20000
         BTREE_SIZE=20000
 
+        COMP="UTEST_vos_nvme"
+        if run_test_filter "sudo -E ${SL_PREFIX}/bin/vos_tests -a"; then
+            cat /proc/meminfo
+            # Setup AIO device
+            AIO_DEV=$(mktemp /tmp/aio_dev_XXXXX)
+            dd if=/dev/zero of="${AIO_DEV}" bs=1G count=4
+
+            # Setup daos_nvme.conf
+            NVME_CONF="/mnt/daos/daos_nvme.conf"
+            cp -f src/vos/tests/daos_nvme.conf ${NVME_CONF}
+            sed -i "s+\"filename\": \".*\"+\"filename\": \"${AIO_DEV}\"+g" ${NVME_CONF}
+
+            export VOS_BDEV_CLASS="AIO"
+            run_test "sudo -E ${SL_PREFIX}/bin/vos_tests" -a
+
+            rm -f "${AIO_DEV}"
+            rm -f "${NVME_CONF}"
+        fi
+
         COMP="UTEST_vos"
-        cat /proc/meminfo
-        # Setup AIO device
-        AIO_DEV=$(mktemp /tmp/aio_dev_XXXXX)
-        dd if=/dev/zero of="${AIO_DEV}" bs=1G count=4
+        run_test src/vos/tests/evt_stress.py
+        run_test src/vos/tests/evt_stress.py --algo dist_even
+        run_test src/vos/tests/evt_stress.py --algo soff
 
-        # Setup daos_nvme.conf
-        NVME_CONF="/mnt/daos/daos_nvme.conf"
-        cp -f src/vos/tests/daos_nvme.conf ${NVME_CONF}
-        sed -i "s+\"filename\": \".*\"+\"filename\": \"${AIO_DEV}\"+g" ${NVME_CONF}
-
-        export VOS_BDEV_CLASS="AIO"
-        run_test "sudo -E ${SL_PREFIX}/bin/vos_tests" -a
-
-        rm -f "${AIO_DEV}"
-        rm -f "${NVME_CONF}"
     else
         if [ "$RUN_TEST_VALGRIND" = "memcheck" ]; then
             [ -z "$VALGRIND_SUPP" ] &&
@@ -129,6 +175,7 @@ if [ -d "/mnt/daos" ]; then
                                           --show-reachable=yes \
                                           --num-callers=20 \
                                           --error-limit=no \
+                                          --fair-sched=try \
                                           --suppressions=${VALGRIND_SUPP} \
                                           --gen-suppressions=all \
                                           --error-exitcode=42 \
@@ -153,10 +200,27 @@ if [ -d "/mnt/daos" ]; then
     COMP="UTEST_vos"
     run_test "${SL_PREFIX}/bin/vos_tests" -A 500
     run_test "${SL_PREFIX}/bin/vos_tests" -n -A 500
+    COMP="UTEST_vos"
+    cmd="-c pool -w key@0-4 key@3-4 -R key@3-3 -w key@5-4 -R key@5-3 -a -i -d -D"
+    run_test "${SL_PREFIX}/bin/vos_tests" -r "\"${cmd}\""
+    cmd="-c pool -w key@0-3 key@3-4 -w key@5-1 -w key@5-4 -R key@5-3 -a -i -d -D"
+    run_test "${SL_PREFIX}/bin/vos_tests" -r "\"${cmd}\""
+    cmd="-c pool -x key@10-400 -i -d -o pool -a -i -d -D"
+    run_test "${SL_PREFIX}/bin/vos_tests" -r "\"${cmd}\""
+    export DAOS_DKEY_PUNCH_PROPAGATE=1
+    run_test "${SL_PREFIX}/bin/vos_tests" -C
+    unset DAOS_DKEY_PUNCH_PROPAGATE
 
     COMP="UTEST_vea"
     run_test "${SL_PREFIX}/bin/vea_ut"
     run_test "${SL_PREFIX}/bin/vea_stress -d 60"
+    # regression test for DAOS-12256
+    COMP="UTEST_vea_debug"
+    export D_LOG_MASK=DEBUG
+    export DD_SUBSYS=all
+    export DD_MASK=all
+    run_test "${SL_PREFIX}/bin/vea_ut"
+    unset D_LOG_MASK DD_SUBSYS DD_MASK
 
     COMP="UTEST_bio"
     run_test "${SL_BUILD_DIR}/src/bio/smd/tests/smd_ut"
@@ -233,13 +297,16 @@ if [ -d "/mnt/daos" ]; then
     # Reporting
     if [ "$failed" -eq 0 ]; then
         # spit out the magic string that the post build script looks for
-        echo "SUCCESS! NO TEST FAILURES"
+        produce_output 0 "SUCCESS! NO TEST FAILURES"
     else
-        echo "FAILURE: $failed tests failed (listed below)"
+        fail_msg="FAILURE: $failed tests failed (listed below)
+"
         for ((i = 0; i < ${#failures[@]}; i++)); do
-            echo "    ${failures[$i]}"
+            fail_msg=$"$fail_msg    ${failures[$i]}
+"
         done
-        if ! ${OLD_CI:-true}; then
+        produce_output 1 "$fail_msg"
+        if ! ${IS_CI:-false}; then
             exit 1
         fi
     fi
